% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{add_colors}
\alias{add_colors}
\title{Add a distinct color to each group}
\usage{
add_colors(node_table, group_colors = NULL)
}
\arguments{
\item{node_table}{A data frame in which rows correspond to the nodes of a
network. Must contain a column names \code{group}.}

\item{group_colors}{Optional (default: \code{NULL}), vector of character strings
representing colors, with a color for each node group. Colors may be
provided either as color names, e.g. 'green', or as hex, e.g. "#00FF00".}
}
\value{
Returns \code{node_table} with 'color' column, that contains a different
color for each node group.
}
\description{
Adds a 'color' column with hex values to \code{node_table}. The colors will be
assigned based on the 'group' column of \code{node_table}. Nodes that are in the
same group obtain the same color.
}
