% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{add_node_pos}
\alias{add_node_pos}
\title{Calculate x and y coordinates for nodes}
\usage{
add_node_pos(node_table, nodesize, layout = c("circle"), space_fct = 1.2)
}
\arguments{
\item{node_table}{A data frame in which rows correspond to the nodes of a
network.}

\item{nodesize}{A number that represents the maximum node width. This will be
used to calculate the minimum radius that a circular layout can have to
avoid overlap of nodes.}

\item{layout}{A character string (default: 'circle') that determines which
layout will be used to arrange the nodes.}

\item{space_fct}{A number (default: 1.2) that is used as multiplier for
nodesize. It increases the spacing between nodes.}
}
\value{
Returns \code{node_table} with 'x' and 'y' column added or overwritten.
}
\description{
Calculates coordinates for each node in the node table. To avoid overlap the
size of nodes is takes into account when creating the layout.
}
