% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{adj_matrix_to_edgelist}
\alias{adj_matrix_to_edgelist}
\title{Adjacency matrix to edge list}
\usage{
adj_matrix_to_edgelist(adj_matrix)
}
\arguments{
\item{adj_matrix}{Data frame or matrix. Square adjacency matrix. Nodes must be specified
in row names and column names.}
}
\value{
Returns data frame representing a network as weighted edge list. Columns are
source, target, and weight.
}
\description{
Converts an adjacency matrix into an weighted edge list. Interactions of nodes
with themselves are removed. Assumes the matrix to be symmetric over the
diagonal, the resulting edges will be undirected.
}
