% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{adj_matrix_to_nodetable}
\alias{adj_matrix_to_nodetable}
\title{Create node table from adjacency matrix}
\usage{
adj_matrix_to_nodetable(adj_matrix)
}
\arguments{
\item{adj_matrix}{An adjacency matrix (data frame or matrix), with column
names.}
}
\value{
Returns a data frame with a single column: 'node', which holds the
names of the network's nodes.
}
\description{
Extracts column names from the adjacency matrix and creates a basic node table
with the column 'node'.
}
