% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anvis-package.R
\docType{package}
\name{anvis}
\alias{anvis}
\title{anvis: A package for convenient network visualizations}
\description{
Here there will be a short package functionality summary later
}
\section{Foo functions}{

The foo functions ...
}

