% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfsFromNet.R
\name{dfsFromNetwork}
\alias{dfsFromNetwork}
\title{Convert igraph or graphNEL networks to list networks}
\usage{
dfsFromNetwork(networks)
}
\arguments{
\item{networks}{Objects of class "igraph" or "graphNEL". Can be a single
object or list.}
}
\value{
Returns a list of two dataframes, named "vertices", and "edges". "vertices"
will contain as first column "node" with the node names, additional
columns will be used for further node attributes. The edges data frame
contains as first two columns "source" and "target", and will contain
further columns for edge attributes. Return will be a list of lists
if multiple networks have been provided as input.
}
\description{
This function takes objects of class "igraph" or "graphNEL" (or a list
of these objects), and converts them into a networks represented as lists
of data frames.
}
\examples{
# We create basic networks for this example
nets <- adjToNetwork(sepsis[1:2],
             node_attrs = "size",
             edge_attrs = "width",
             width_type = "partcor")

# Convert networks
dfsFromNetwork(nets)
}
