% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{edge_weight_to_widths}
\alias{edge_weight_to_widths}
\title{Calculate edge widths}
\usage{
edge_weight_to_widths(
  edge_table,
  width_type = c("default_scaling", "MI", "cor", "partcor", "ranked", "percentile")
)
}
\arguments{
\item{edge_table}{A data frame with a column called "weight" containing
numeric values.}

\item{width_type}{A character string that determines the method to be used for
converting edge weight to edge width. Options are: \code{"cor"}, which is
intended to be used with Pearson or Spearman correlation values (range -1
to 1). Widths will be the absolute value of the correlation, scaled with a
sigmoid. \code{"partcor"}, which is intended to be use for partial correlation
values (range -1 to 1). For this method, edge widths will be the cube root
of absolute weight values, scaled with a sigmoid. The option \code{"MI"} is meant
to be used with weights derived from mutual information (range 0 to +∞).
Widths will be the weights divided by to maximum weight, then scaled with
a sigmoid. \code{"default_scaling"} applies the same transformation as \code{"MI"},
as a result scaling the any weights to range 0 to 1.
\code{"ranked"} will calculate percentage ranks of the weight, and
scale them with a sigmoid. \code{"percentile"} will chose a set of fixed widths
determined by the percentile of a weight value. The highest percentiles will
be assigned the largest width, but they are the smallest group, vice versa
for the lowest percentiles. This argument can be abbreviated.}
}
\value{
Return the input data frame with an added column "width", which can
be used to scale edges in a network visualization.
}
\description{
Adds a column called "width" to the input \code{edge_table}. The widths for the
edges can be calculated using five different methods. Which method is most
appropriate depend on the range of the input data. For more details on the
choice of the method check \code{width_type} in the Arguments section.
}
\details{
The methods have in common that they relate to the absolute values of the column
"weight", which is a required argument. The most extreme values for "weight"
get the highest value for edge width. The range of edge widths will be between
0 and 1, irrespective of the chosen method.
}
