% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{edgelist_to_adj}
\alias{edgelist_to_adj}
\title{Convert edge list to adjacency matrix}
\usage{
edgelist_to_adj(edge_list, weight_col = "weight")
}
\arguments{
\item{edge_list}{Data frame. The first two columns should describe the
connected nodes. Another column with edge weights can be used by this
function. Any additional columns/attributes of the edges will be lost.}

\item{weight_col}{Character string, the name of the column that contains edge
weights (default: 'weight'). If this argument is NULL all edges will have
weight 1 in the adjacency matrix.}
}
\value{
Returns a square adjacency matrix with column and row names representing
the interacting nodes. The matrix will by symmetric along the diagonal.
}
\description{
Creates an adjacency matrix from edgelist.
}
