% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_igraph.R
\name{get_param}
\alias{get_param}
\title{Use user parameters over default}
\usage{
get_param(graph, plot_params, plot_arg, name_in_df = NULL)
}
\arguments{
\item{graph}{Igraph object.}

\item{plot_params}{Names list with the graphical parameters supplied by the
user.}

\item{plot_arg}{Character string. Will be used to modify visualization with
\code{igraph::plot.igraph}.}

\item{name_in_df}{Character string. Some attributes are called different in the
context of plotting vs in the object they are stored in (e.g. vertex.label
vs name, respectively)}
}
\value{
A vector of values that will be used for one of the arguments of
\code{igraph::plot.igraph}. When the attribute is not in
}
\description{
This is a helper function for \code{vis_igraph}. I makes sure that when there are
plotting arguments that have been provided by the user, these will be used
instead of the default found in the graph.
}
