% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{group_nodes}
\alias{group_nodes}
\title{Add group column and sort node table}
\usage{
group_nodes(node_table, group_vec)
}
\arguments{
\item{node_table}{A data frame in which rows correspond to the nodes of a
network.}

\item{group_vec}{A vector contains group labels for each row in \code{node_table}.}
}
\value{
Returns \code{node_table} with 'group' column added, sorted by group.
}
\description{
Adds 'group' column to \code{node_table} with values corresponding to  \code{group_vec}.
The rows of \code{node_table} will be reordered so the group column is sorted
alphabetically.
}
