% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{node_size_connectivity}
\alias{node_size_connectivity}
\title{Add connectivity based 'size' column to node table}
\usage{
node_size_connectivity(
  node_table,
  edge_table,
  size_type = c("igraph", "cytoscape", "scaled_only")
)
}
\arguments{
\item{node_table}{A data frame with a column named 'node'.}

\item{edge_table}{A data frame with with columns source and target, containing
the names of nodes connected by an edge, and a column called "weight" containing
numeric values.}

\item{size_type}{One of "igraph" (default), "cytoscape", or "scaled_only".
This argument determines a factor for linear scaling of node size. Factors
are 15, 25, and 1, respectively.}
}
\value{
Returns \code{node_table} with an added column 'size'.
}
\description{
Calculates the connectivity of nodes as the sum of all weights of edges
originating or ending at a node.
For use as node size, the values are scaled to range 0 to 1, transformed
with a sigmoid, and then multiplied with a constant factor that is determined
by \code{size_type}.
}
