% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paleo_mi}
\alias{paleo_mi}
\title{Mutual information matrix for elements detected in deep sea drilling}
\format{
\subsection{\code{paleo_mi}}{

A matrix made of 38 rows and columns. The rows and columns represent
several trace elements and depth and age of the ground. The cells of the matrix
give mutual information values that give an estimate for the extent to which
the abundance of the elements show the same patterns.
}
}
\source{
For more information on the raw data or a download thereof:
\url{doi:10.1016/j.margeo.2005.02.009}
}
\usage{
paleo_mi
}
\description{
A dataset containing mutual information (MI) values, that indicate which trace
elements show the same trends in the samples of a deep sea drilling experiment.
The MI data presented here was generated from element abundance data that has
previously been investigated to study the environmental change that
occurred late Paleocene–early Eocene.
}
\keyword{datasets}
