% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{percentile_widths}
\alias{percentile_widths}
\title{Edge width based on percentiles}
\usage{
percentile_widths(n_edges)
}
\arguments{
\item{n_edges}{Integer to determine the number of edge widths that should be
returned}
}
\value{
A vector of edge widths which matches \code{n_edges} with its length. Widths
vary between 1 and 0.025.
}
\description{
Creates a vector of edge widths that matched with the number of edges. Edges
widths are determined on the percentiles of connection strengths. The highest
2 percentiles have width 10, the next 3 percentiles have width 8, etc.
}
