% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sepsis}
\alias{sepsis}
\title{Partial correlation between the levels of 36 proteins measured in blood plasma}
\format{
\subsection{\code{sepsis}}{

A list of 12 named elements that relate to different groups of hospital
patients. Each list element corresponds to a matrix with
36 rows and 36 columns, representing the 36 proteins of which the concentration
was measured. Each cell in the matrix shows the partial correlation between
the levels of proteins measured in blood plasma. The matrices are symmetrical
over the diagonal.

List elements were made by creating subsets of hospital patients, for which
partial correlation was estimated separately. The names of the list element
and the subgroup of patients they relate to is as follows (note that NSTI
stands for necrotizing soft tissue infection):

\describe{
\item{Hospital patients}{NSTI patients}
\item{Control group 1}{Non NSTI patients}
\item{Control group 2}{Cellulitis patients}
\item{Surgical control}{Surgical control patients}
\item{Septicshock patients}{NSTI patients who had septicshock}
\item{No-septicshock patients}{NSTI patients who didn't have septicshock}
\item{Mono-microbial patients}{NSTI caused by 1 bacterial species (mono-microbial)}
\item{Poly-microbial patients}{NSTI caused by several species (poly-microbial)}
\item{Mono-microbial septicshock patients}{Mono-microbial patients with septicshock}
\item{Mono-microbial no-septicshock patients}{Mono-microbial patients without septicshock}
\item{Poly-microbial septicshock patients}{Poly-microbial patients with septicshock}
\item{Poly-microbial no-septicshock patients}{Poly-microbial patients without septicshock}
}
}
}
\source{
For more information on the raw data or a download thereof:
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.f1vhhmgw4}
}
\usage{
sepsis
}
\description{
A dataset containing partial correlations values, that indicate how similar the
levels of selected blood plasma proteins behave. The protein levels were
measured in a range of hospital patients, which can be assigned to different
groups based on the type of their disease.
}
\keyword{datasets}
