% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{sort_avg_connectivity}
\alias{sort_avg_connectivity}
\title{Reorder node tables by their average connectivity}
\usage{
sort_avg_connectivity(nodes_list)
}
\arguments{
\item{nodes_list}{A list of node tables. Each must have a 'node' column holding
node names. A least some of the tables should have a 'size' column, containing
the connectivity information.}
}
\value{
Returns the same \code{nodes_list} as was used as input, with rows reordered.
}
\description{
Calculates the average connectivity and reorders the rows of the input node
tables. If a 'group' column is present in the first node table, the groups
will only be ordered internally by average connectivity.
}
