% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_igraph.R
\name{vis_igraph}
\alias{vis_igraph}
\title{Visualize with igraph}
\usage{
vis_igraph(
  edge_table = NULL,
  node_table = NULL,
  igraph_obj = NULL,
  radial_labs = T,
  rad_lab_opts = list(),
  scale_width = 3.25,
  ...
)
}
\arguments{
\item{radial_labs}{A logical to indicate whether vertex labels should be
positioned radially around the circular arrangement of vertices.}

\item{rad_lab_opts}{A names list, in which the names are valid arguments for
\code{text()}.}

\item{...}{Additional options to be used with \code{igraph::plot.igraph} for
visualizing your network. Any options provided here will overwrite the
defaults. If \code{radial_labs} is \code{FALSE}, this argument can also be used to
customize the vertex labels placed by \code{plot.igraph}.}
}
\value{
The section on the returned values
}
\description{
More extensive description of what the function does
}
\details{
When all three network input parameters are provided, \code{edge_table} and
\code{node_table} will be used instead of \code{igraph_obj}.
Used defaults in the for vertex and label styling, but these can be overwritten
by providing arguments for igraph plot style, i.e. arguments starting
with the "vertex." or "label.". See igraph plotting docs.
}
