% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cytoscape.R
\name{visCytoscape}
\alias{visCytoscape}
\title{Visualize in Cytoscape}
\usage{
visCytoscape(
  network,
  directed = if (is(network, "graphNEL")) {
     graph::edgemode(network) ==
    "directed"
 } else if (is(network, "igraph")) {
     igraph::is_directed(network)
 }
    else {
     FALSE
 },
  radial_labs = TRUE,
  export_image = TRUE,
  save_session = TRUE,
  close_session = export_image,
  save_name = "network",
  export_type = c("PNG", "JPEG", "PDF", "SVG", "PS"),
  export_opts = list(),
  scale_width = 1,
  node_space = 1.2,
  cyto3.8_check = T
)
}
\arguments{
\item{network}{A network object of one of the valid classes:
"igraph", "graphNEL", and "list" of two data frames with the names "edges"
and "vertices". If the nodes have an additional attribute 'color' or
'size', the respective the respective element will be adjusted in the
visualization. If edges have the attributes 'width' or 'weight', this will
be used to determine edge widths in the visualization (only when 'width'
is not present weight will be used). Edge attribute 'color' determines
the color of edges.}

\item{directed}{A logical (automatically determined for network
objects of class 'igraph' and 'graphNEL', otherwise default \code{FALSE}).
When this argument is \code{TRUE} edges will be drawn with arrows.}

\item{radial_labs}{A logical (default \code{TRUE}) to indicate whether node labels
should be positioned radially around the circular arrangement of nodes.}

\item{export_image}{Logical (default \code{TRUE}), should the visualization be saved?}

\item{save_session}{Logical (default \code{TRUE}), specifying whether the cytoscape
session should be saved (as .cys file).}

\item{close_session}{Logical (default \code{TRUE}), should the cytoscape session be
closed after optional saving?}

\item{save_name}{Character string (default 'network') that will be used to
name image and session save files. File extensions should not be included
for this argument, as they will be added automatically. If the name
chosen for \code{save_name} already exists in the current working directory
numbers will be appended to it.}

\item{export_type}{Character string, one of 'PNG' (default), 'JPEG', 'PDF',
'SVG', or 'PS'. Determines which file format will be used to save the
network visualization.}

\item{export_opts}{List with named values that will be used to customize
image export. Any argument accepted by \link[RCy3:exportImage]{RCy3::exportImage} is valid.
The options 'filename' and 'type' of this list will be overwritten
by \code{save_name} and \code{export_type}, respectively.}

\item{scale_width}{Numeric, a number that will be multiplied with the edge
widths, scaling the edge widths linearly.}

\item{node_space}{Numeric (default 1.2). Adjusts spacing between nodes.
When this argument is 1, the borders of the largest nodes
can be touching. When e.g. a value of 2 is chosen for this argument, the
space around a node will be twice the (maximum) node size.}

\item{cyto3.8_check}{Logical (default \code{TRUE}). Should execution stop if
Cytoscape version 3.8.x is detected? \code{FALSE} to skip this test.
Cytoscape version 3.8.x has problems interacting with \code{RCy3}, first
visualization may not show.}
}
\description{
Uses RCy3 to communicate to your running cytoscape software. Visualizes your
network with a circular layout. Depending on the attributes of your
network, additional visual properties will be added to
the visualization. Image and session object can be saved.
}
