% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{weights_to_color}
\alias{weights_to_color}
\title{Color edges based on weight}
\usage{
weights_to_color(edge_table)
}
\arguments{
\item{edge_table}{A data frame with a column called "weight" containing
numeric values.}
}
\value{
Returns the input \code{edge_table} with a column called 'color' added
(or overwritten if it was already present).
}
\description{
This function adds a column of hex colors to and edge table. The colors are
determined by the weight of an edge. When both positive and negative weights
are found, a blue-red divergent color palette will be used to select colors.
For only positive weights, colors are selected from a white to red color
gradient.
}
\details{
Colors are assigned by placing the edge weights into 100 bins, and assigning
100 colors from the appropriate palette.
}
