% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{adjToNetwork}
\alias{adjToNetwork}
\title{Create edge and node table from adjacency matrix}
\usage{
adjToNetwork(
  adj_matrix,
  directed = FALSE,
  self_loops = FALSE,
  node_attrs = c("none", "all", "group", "color_group", "size"),
  edge_attrs = c("none", "all", "width", "color"),
  group_vec = NULL,
  group_colors = NULL,
  size_type = NULL,
  width_type = NULL,
  edge_color_func = NULL
)
}
\arguments{
\item{adj_matrix}{Data frame or matrix. Square adjacency matrix. Nodes must be
specified in row names and column names.}

\item{directed}{Logical (default \code{FALSE}), whether edges are directed in the
network. If \code{FALSE}, the information in the lower triangle of the
adjacency matrix will be discarded.}

\item{self_loops}{Logical (default \code{FALSE}). Should the values for interaction
of nodes with themselves (diagonal of the adjacency matrix) be retained?}

\item{node_attrs}{Character strings, one or multiple of 'none', 'all', 'group',
'color_group', and 'size'. This argument can be used to choose which
additional attributes should be added to the node table. Selecting 'group'
will append a column to the node table that corresponds to \code{group_vec}.
Selecting 'color_group' in addition to 'group' will add a column that
contains a color for each node, selected based on the group of this node.
Selecting 'size' will add a column with node sizes, that are based on the
connectivity of the nodes.}

\item{edge_attrs}{Character strings, one or multiple of 'none, 'all', 'width',
'color'. This argument can be used to choose which additional attributes
should be added to the edge table. Selecting 'width' will add a column to
edge table that contains edge width values. These are determined by scaling
the edge weights to range 0 to 1 using a sigmoid and a method based on
\code{width_type} (more info in \link{edge_weight_to_widths}).}

\item{group_vec}{A vector contains group labels for each row in \code{node_table}.}

\item{group_colors}{Optional (default: \code{NULL}), vector of character strings
representing colors, with a color for each node group. Colors may be
provided either as color names, e.g. 'green', or as hex, e.g. "#00FF00".}

\item{size_type}{One of "igraph" (default), "cytoscape", or "scaled_only".
This argument determines a factor for linear scaling of node size. Factors
are 15, 25, and 1, respectively.}

\item{width_type}{A character string that determines the method to be used for
converting edge weight to edge width. Options are: \code{"cor"}, which is
intended to be used with Pearson or Spearman correlation values (range -1
to 1). Widths will be the absolute value of the correlation, scaled with a
sigmoid. \code{"partcor"}, which is intended to be use for partial correlation
values (range -1 to 1). For this method, edge widths will be the cube root
of absolute weight values, scaled with a sigmoid. The option \code{"MI"} is meant
to be used with weights derived from mutual information (range 0 to +∞).
Widths will be the weights divided by to maximum weight, then scaled with
a sigmoid. \code{"default_scaling"} applies the same transformation as \code{"MI"},
as a result scaling the any weights to range 0 to 1.
\code{"ranked"} will calculate percentage ranks of the weight, and
scale them with a sigmoid. \code{"percentile"} will chose a set of fixed widths
determined by the percentile of a weight value. The highest percentiles will
be assigned the largest width, but they are the smallest group, vice versa
for the lowest percentiles. This argument can be abbreviated.}

\item{edge_color_func}{A function that takes as first argument a number, and
returns a vector of colors, the length of which equals the input number.
The first colors in this vector will be used for low values, the last
colors will be assigned to high values. This argument is optional
(default \code{NULL}).}
}
\value{
Will return a list with two data frames, named 'edge_table' and
'node_table'. The 'node_table' data frame has a 'node' column and other
columns for the additional attributes that were added. The 'edge_table' has
columns 'source' and 'target', and more columns for the added attributes.
}
\description{
Creates edge and node tables with various optional attributes.
}
\details{
When \code{node_attrs} or \code{edge_attrs} is a vector containing 'none' and any other
option, no additional attributes will be added.

When selecting 'group' as node attribute \code{group_vec} becomes a required
argument. When selecting 'color_group' as a node attribute, 'group' must also
be selected. The colors for the groups can be adjusted with \code{group_colors},
but this is optional. When selecting 'size' as node attribute the range of
node sizes can be adjusted with the \code{size_type} argument (when this argument
is NULL the nodes sizes will be set to work neatly for igraph visualization).
When 'width' is added as edge attribute, it is wise to check the options for
\code{width_type}, to obtain width that show the most logical contrasts for your
data. \code{width_type = NULL} (default), will scale the data to range 0 to 1 and
apply a sigmoid, so the lowest edge weights will have even lower width, and
higher weight maintain high values.
}
