% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_wrappers.R
\name{anvis}
\alias{anvis}
\title{Visualize or export networks}
\usage{
anvis(
  networks,
  directed = NULL,
  save_names = "network",
  output_type = c("igraph", "cytoscape", "network"),
  vis_edge_factor = NULL,
  vis_radial_labs = T,
  vis_save = FALSE,
  vis_export_type = c("png", "jpeg", "pdf", "svg", "ps"),
  vis_export_opts = list(),
  igr_rad_lab_opts = list(),
  igr_plot_opts = list(),
  igr_par_opts = list(),
  igr_grid = FALSE,
  igr_grid_names = FALSE,
  cyto3.8_check = TRUE,
  cyto_save_session = FALSE,
  cyto_close_session = vis_save,
  cyto_node_space = 1.2,
  netw_ext = c("XGMML", "table", "sif", "tab", "tgf", "net"),
  netw_xgmml_title = NULL
)
}
\arguments{
\item{networks}{A network or a list of these. Valid classes for the network are
"igraph", "graphNEL", and "list" of two data frames with the names "edges"
and "vertices". If the edges in the network have the additional attributes
'width' or 'color' those attributes will be used to style the edges in
the visualization. Optional node attribute 'group' will be used to arrange
vertices so those with the same group label will be placed next to each
other, 'color' and 'size' will be used to adjust the corresponding
features of the vertices.}

\item{directed}{Logical (default \code{NULL}), whether visualizations should
feature edge curvature and arrows. If this argument is \code{NULL} this
property will be automatically determined for 'igraph' or 'graphNEL'
object, and otherwise by it will be \code{FALSE} by default.}

\item{save_names}{Base name for the files that are saved by this function
(default: 'network'). If a single name is provided for multiple networks,
numbers will be appended to the base name while saving. Alternatively,
this argument can be a vector with a name for each network that will be
saved. If in any case one of the same names already exist in the save
target directory, numbers will added to the base file name, and incremented
until a no file with the same base name and number exists.}

\item{output_type}{Choose 'igraph' (default), 'cytoscape', 'network', or
'return_only', can be abbreviated. This argument determines the main
output of this function. The first two options will make a
visualization with the respective software. The third option will save
the networks in a format that is determined by the \code{netw_ext} argument.
The 'return_only' returns the networks created by this function, just
like all other options, but without any side effects.}

\item{vis_edge_factor}{Numeric, a number that will be multiplied with the edge
widths, scaling the edge widths linearly. When this argument is \code{NULL},
default values will be used (3.25 for igraph and 2 for cytoscape).}

\item{vis_radial_labs}{A logical (default \code{TRUE}) to indicate whether node
labels should be positioned radially around the circular arrangement of
vertices. If \code{FALSE}, they will be placed on top or next to the vertices,
depending on your visualization method. This argument is only used when
the output type is 'cytoscape' or 'igraph'.}

\item{vis_save}{Logical (default \code{FALSE}), should network visualizations be saved?
If this parameter is \code{FALSE} igraph will show a plot in your R session and
cytoscape will keep the session with all networks open.}

\item{vis_export_type}{Character string, one of 'png' (default), 'jpeg', 'pdf',
'svg' and 'ps'. Use this argument to choose the file type with which
visualizations are saved.}

\item{vis_export_opts}{A list with named elements. The list items will be used
as arguments for the function that exports your visualization. When
exporting a visualization made with igraph, the available options for
exporting are those that work for the graphical device you selected with
\code{vis_export_type}. In this case, you can check which options are available
for your graphical device with e.g. \link{png}. When the visualization was
made with cytoscape, valid export options are those accepted by
\link[RCy3:exportImage]{RCy3::exportImage}.}

\item{igr_rad_lab_opts}{A named list, in which the names are valid arguments
for \link{text}. These styling options apply to igraph node labels when
\code{vis_radial_labs} is \code{TRUE}.}

\item{igr_plot_opts}{A named list containing additional options to be used with
\link[igraph:plot.igraph]{igraph::plot.igraph} for customizing the igraph visualizations. If
\code{vis_radial_labs} is \code{FALSE}, this argument can also be used to
customize igraph visualization node labels.}

\item{igr_par_opts}{A list with named elements. The list items will be used
as arguments for \link{par} while making visualizations with igraph. Changes to
the graphical parameters will be applied to your selected graphical device
before making the visualization. After the visualization is completed,
the graphical parameters will be reset to their original value.}

\item{igr_grid}{Logical (default: \code{FALSE}), or a vector with two numbers
representing grid dimensions. This argument will determine whether igraph
visualization are be made separately, or arranged on a grid in a single
image. \code{TRUE} will automatically determine grid size.}

\item{igr_grid_names}{Logical (default: \code{FALSE}) determining whether titles
should be placed above the individual networks in the grid layout. \code{TRUE}
will use the names of the \code{adj_mats} list to place titles. Alternatively,
a vector with a name for each network can be provided as input for this
argument.}

\item{cyto3.8_check}{Logical (default \code{TRUE}). Should execution stop if
Cytoscape version 3.8.x is detected? \code{FALSE} to skip this test.
Cytoscape version 3.8.x has problems interacting with \code{RCy3}, first
visualization may not show.}

\item{cyto_save_session}{Logical (default \code{FALSE}). When visualizing with
cytoscape, should the session be saved (as .cys file)?}

\item{cyto_close_session}{Logical (default same as \code{vis_save}), should the
cytoscape session be closed after visualization has been completed?}

\item{cyto_node_space}{Numeric (default 1.2). By default nodes are arranged
with a circular layout, where the space between the nodes is adjusted with
this argument. When this argument is 1, the borders of the largest nodes
can be touching. When e.g. a value of 2 is chosen for this argument, the
space around a node will be twice the (maximum) node size. This argument
only increases node spacing for cytoscape visualizations.}

\item{netw_ext}{Character string (default 'XGMML'). When \code{output_type} is
'network', which file format should be used to save the network? Options
are 'XGMML', 'table', 'sif', 'tab', 'tgf', 'net'.}

\item{netw_xgmml_title}{Only used when \code{output_type} is 'network' and \code{netw_ext}
is 'XGMML'. This argument determines the title attribute of a network
when it's saved as XGMML file. Possible values are \code{NULL} (default), to use
the \code{save_names} as a network title, a single string, to give all networks
the same title, or a vector of strings matching the length of input
networks, to provide each networks its own title.}
}
\description{
This function takes a (list of) network object(s) and visualizes them or
exports them in various formats. Based on additional attributes of
nodes and edges, visual elements such as edge widths and node sizes are
adjusted in the visualizations. Networks can be visualized with Cytoscape
or igraph. For the latter, there are options available to visualize multiple
networks in a grid, for easier comparison.
}
\details{
Note: When using this function to create visualizations with Cytoscape, the
cytoscape software needs to be running in the background already.
}
\seealso{
Generate more interesting visualizations by using \link{addVisAttrs}
for adding attributes to an existing network objects, and
\link{adjToNetwork} for creating a network from an adjacency matrix and
adding node and edge attributes in a single step.
}
