% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_styles.R
\name{n_distinct_cols}
\alias{n_distinct_cols}
\title{Select n colors of distinct appearance}
\usage{
n_distinct_cols(n, colors = NULL)
}
\arguments{
\item{n}{Single numeric value that determines how many colors are returned.}
}
\value{
A character vector containing the number of requested colors
}
\description{
This function contains a pre-selected collection of colors and will provide
the selected number of these. Optionally, custom vectors with colors can be
provided. If the number of requested colors exceeds the avaible colors (25),
the available colors will be repeated.
}
\examples{
n_distinct_cols(6)
}
