% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_manipulation.R
\name{node_size_connectivity}
\alias{node_size_connectivity}
\title{Add connectivity based 'size' column to node table}
\usage{
node_size_connectivity(
  node_table,
  adj_matrix,
  size_type = c("igraph", "cytoscape", "scaled_only")
)
}
\arguments{
\item{node_table}{A data frame with a column named 'node'. The values in this
column (names of nodes), must be the same as the row names
of \code{adj_matrix}. Having nodes in the same order in \code{node_table} and the
matrix not required.}

\item{adj_matrix}{Data frame or matrix. Square adjacency matrix. Nodes must be specified
in row names and column names.}

\item{size_type}{One of "igraph" (default), "cytoscape", or "scaled_only".
This argument determines a factor for linear scaling of node size. Factors
are 15, 25, and 1, respectively.}
}
\value{
Returns \code{node_table} with an added column 'size'.
}
\description{
Calculates the connectivity of nodes as the row sums from an adjacency matrix.
For use as node size, the values are scaled to range 0 to 1, transformed
with a sigmoid, and then multiplied with a constant factor that is determined
by \code{size_type}
}
