% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cytoscape.R
\name{vis_in_cytoscape}
\alias{vis_in_cytoscape}
\title{Visualize in Cytoscape}
\usage{
vis_in_cytoscape(
  node_table,
  edge_table,
  radial_labs = TRUE,
  export_image = TRUE,
  save_session = TRUE,
  close_session = TRUE,
  save_name = "network",
  export_type = c("PNG", "JPEG", "PDF", "SVG", "PS"),
  export_opts = list(),
  scale_width = 1,
  node_space = 1.2,
  cyto3.8_check = T
)
}
\arguments{
\item{node_table}{Data frame with required column 'node'. Optional are columns
'color' and 'size': when present they will be used to determine the respective
visual properties of the visualization. Other columns with node attributes
will not be used in the visualization, but they will be added to the node
table in cytoscape.}

\item{edge_table}{Data frame with required columns 'source' and 'target'.
Optional are columns: 'width' or 'weight' to determine edge widths in the
visualization (only when 'width' is not present weight will be used), and
'color' to determine the color of the edges.}

\item{export_image}{Logical (default \code{TRUE}), should the visualization be saved?}

\item{save_session}{Logical (default \code{TRUE}), specifying whether the cytoscape
session should be saved (as .cys file).}

\item{close_session}{Logical (default \code{TRUE}), should the cytoscape session be
closed after optional saving?}

\item{save_name}{Character string that will be used to name image and session
save files, excluding file extensions, as these will be added automatically.
If this argument is \code{NULL}, \code{image_opt[["filename"]]} will be used for naming.
If both are \code{NULL}, "network" will be used as default name. If the name
chosen for \code{save_name} already exists in the current working directory
numbers will be appended to it.}

\item{export_opts}{List with named values that will be used to customize
image export. Any argument accepted by \link[RCy3:exportImage]{RCy3::exportImage} is valid.}

\item{cyto3.8_check}{Logical (default \code{TRUE}). Should execution stop if
Cytoscape version 3.8.x is detected? \code{FALSE} to skip this test.
Cytoscape version 3.8.x has problems interacting with \code{RCy3}, first
visualization may not show.}
}
\description{
Uses RCy3 to communicate to your running cytoscape software. Visualizes your
network with a circular layout. Depending on the attributes of your
\code{edge_table} and \code{node_table}, additional visual properties will be added to
the visualization. Image and session object can be saved.
}
