% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_wrappers.R
\name{VisualiseNetwork}
\alias{VisualiseNetwork}
\title{Create network and visualize or export}
\usage{
VisualiseNetwork(
  adj_mats,
  node_attrs = c("none", "all", "group", "color_group", "size"),
  edge_attrs = c("none", "all", "width", "color"),
  group_vec = NULL,
  group_colors = NULL,
  width_type = NULL,
  edge_factor = NULL,
  arrange_co = FALSE,
  save_names = "network",
  output_type = c("igraph", "cytoscape", "network", "return_only"),
  vis_radial_labs = T,
  vis_save = FALSE,
  vis_export_type = c("png", "jpeg", "pdf", "svg", "ps"),
  vis_export_opts = list(),
  igr_rad_lab_opts = list(),
  igr_plot_opts = list(),
  igr_par_opts = list(),
  igr_grid = FALSE,
  igr_grid_names = FALSE,
  cyto3.8_check = TRUE,
  cyto_save_session = FALSE,
  cyto_close_session = vis_save,
  cyto_node_space = 1.2,
  netw_ext = c("XGMML", "table", "sif", "tab", "tgf", "net"),
  netw_xgmml_title = NULL
)
}
\arguments{
\item{adj_mats}{A square adjacency matrix or data frame or a list of these.
The data in the matrix is used as edge weights for the network. Row names
and column names specify interacting nodes, and are required.}

\item{node_attrs}{Character strings, one or multiple of 'none', 'all', 'group',
'color_group', and 'size'. This argument can be used to choose which
additional attributes should be added to the node table. Selecting 'group'
will append a column to the node table that corresponds to \code{group_vec}.
Selecting 'color_group' in addition to 'group' will add a column that
contains a color for each node, selected based on the group of this node.
Selecting 'size' will add a column with node sizes, that are based on the
connectivity of the nodes.}

\item{edge_attrs}{Character strings, one or multiple of 'none, 'all', 'width',
'color'. This argument can be used to choose which additional attributes
should be added to the edge table. Selecting 'width' will add a column to
edge table that contains edge width values. These are determined by scaling
the edge weights to range 0 to 1 using a sigmoid and a method based on
\code{width_type} (more info in \link{edge_weight_to_widths}).}

\item{group_vec}{A vector of character strings that assigns group labels to
the nodes. The order of this vector should match the order of column and
rownames of the input adjacency matrices. If \code{adj_mats} is a list, a single
group vector can be used if it matches all adjacency matrices.
Alternatively, provide a list of group vectors with one vector for each
adj. matrix in the list. For this information to be added, \code{node_attrs}
must be 'group' or 'all'.}

\item{group_colors}{Optional (default: \code{NULL}), vector of character strings
representing colors, with a color for each node group. Colors may be
provided either as color names, e.g. 'green', or as hex, e.g. "#00FF00".}

\item{width_type}{Argument used to convert edge weights into widths for
visualizations that are more easy to interpret. Find a description of the
options in \link{edge_weight_to_widths}. This argument can be a single type,
that will be used for all networks, or a vector of types that matches the
number of networks.}

\item{edge_factor}{Numeric, a number that will be multiplied with the edge
widths, scaling the edge widths linearly. When this argument is \code{NULL},
default values will be used (3.25 for igraph and 2 for cytoscape).}

\item{arrange_co}{Logical (default \code{FALSE}), should nodes be reordered based on
their average connectivity in multiple networks? Requires
the same names nodes to be present in all networks. Also requires 'size'
column to be present in node tables, so \code{node_attrs} should be 'all' or
include 'size'.}

\item{save_names}{Base name for the files that are saved by this function
(default: 'network'). If a single name is provided for multiple networks,
numbers will be appended to the base name while saving. Alternatively,
this argument can be a vector with a name for each network that will be
saved. If in any case one of the same names already exist in the save
target directory, numbers will added to the base file name, and incremented
until a no file with the same base name and number exists.}

\item{output_type}{Choose 'igraph' (default), 'cytoscape', 'network', or
'return_only', can be abbreviated. This argument determines the main
output of this function. The first two options will make a
visualization with the respective software. The third option will save
the networks in a format that is determined by the \code{netw_ext} argument.
The 'return_only' returns the networks created by this function, just
like all other options, but without any side effects.}

\item{vis_radial_labs}{A logical (default \code{TRUE}) to indicate whether node
labels should be positioned radially around the circular arrangement of
vertices. If \code{FALSE}, they will be placed on top or next to the vertices,
depending on your visualization method. This argument is only used when
the output type is 'cytoscape' or 'igraph'.}

\item{vis_save}{Logical (default \code{FALSE}), should network visualizations be saved?
If this parameter is \code{FALSE} igraph will show a plot in your R session and
cytoscape will keep the session with all networks open.}

\item{vis_export_type}{Character string, one of 'png' (default), 'jpeg', 'pdf',
'svg' and 'ps'. Use this argument to choose the file type with which
visualizations are saved.}

\item{vis_export_opts}{A list with named elements. The list items will be used
as arguments for the function that exports your visualization. When
exporting a visualization made with igraph, the available options for
exporting are those that work for the graphical device you selected with
\code{vis_export_type}. In this case, you can check which options are available
for your graphical device with e.g. \link{png}. When the visualization was
made with cytoscape, valid export options are those accepted by
\link[RCy3:exportImage]{RCy3::exportImage}.}

\item{igr_rad_lab_opts}{A named list, in which the names are valid arguments
for \link{text}. These styling options apply to igraph node labels when
\code{vis_radial_labs} is \code{TRUE}.}

\item{igr_plot_opts}{A named list containing additional options to be used with
\link[igraph:plot.igraph]{igraph::plot.igraph} for customizing the igraph visualizations. If
\code{vis_radial_labs} is \code{FALSE}, this argument can also be used to
customize igraph visualization node labels.}

\item{igr_par_opts}{A list with named elements. The list items will be used
as arguments for \link{par} while making visualizations with igraph. Changes to
the graphical parameters will be applied to your selected graphical device
before making the visualization. After the visualization is completed,
the graphical parameters will be reset to their original value.}

\item{igr_grid}{Logical (default: \code{FALSE}), or a vector with two numbers
representing grid dimensions. This argument will determine whether igraph
visualization are be made separately, or arranged on a grid in a single
image. \code{TRUE} will automatically determine grid size.}

\item{igr_grid_names}{Logical (default: \code{FALSE}) determining whether titles
should be placed above the individual networks in the grid layout. \code{TRUE}
will use the names of the \code{adj_mats} list to place titles. Alternatively,
a vector with a name for each network can be provided as input for this
argument.}

\item{cyto3.8_check}{Logical (default \code{TRUE}). Should execution stop if
Cytoscape version 3.8.x is detected? \code{FALSE} to skip this test.
Cytoscape version 3.8.x has problems interacting with \code{RCy3}, first
visualization may not show.}

\item{cyto_save_session}{Logical (default \code{FALSE}). When visualizing with
cytoscape, should the session be saved (as .cys file)?}

\item{cyto_close_session}{Logical (default same as \code{vis_save}), should the
cytoscape session be closed after visualization has been completed?}

\item{cyto_node_space}{Numeric (default 1.2). By default nodes are arranged
with a circular layout, where the space between the nodes is adjusted with
this argument. When this argument is 1, the borders of the largest nodes
can be touching. When e.g. a value of 2 is chosen for this argument, the
space around a node will be twice the (maximum) node size. This argument
only increases node spacing for cytoscape visualizations.}

\item{netw_ext}{Character string (default 'XGMML'). When \code{output_type} is
'network', which file format should be used to save the network? Options
are 'XGMML', 'table', 'sif', 'tab', 'tgf', 'net'.}

\item{netw_xgmml_title}{Only used when \code{output_type} is 'network' and \code{netw_ext}
is 'XGMML'. This argument determines the title attribute of a network
when it's saved as XGMML file. Possible values are \code{NULL} (default), to use
the \code{save_names} as a network title, a single string, to give all networks
the same title, or a vector of strings matching the length of input
networks, to provide each networks its own title.}
}
\value{
The main purpose of this function is to create visualizations or
export networks. In addition, this function returns a list of two lists,
containing the information of the created networks. The first is named 'nodes'
and it is a list containing a node table (data frame) for each input adjacency
matrix. The second list is named 'edges' and it contains edge tables. The
return is invisible, so it will not print when not assigned.
}
\description{
This function takes one or more adjacency matrices as input. It uses the
adjacency information (weights) to create several attributes for network
visualization such as scaled edge widths and node sizes based on node
connectivity. The created network with additional attributes can be visualized
with igraph or cytoscape, or exported in several common file formats.
}
\details{
Note: When using this function to create visualizations with cytoscape, the
cytoscape software needs to be running in the background already.
}
