% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_wrappers.R
\name{VisualiseNetwork}
\alias{VisualiseNetwork}
\title{Create formatted circular network}
\usage{
VisualiseNetwork(
  df_adjacency,
  group_vec = NULL,
  vis_type = c("igraph", "cytoscape"),
  width_type = NULL,
  do_save = T,
  save_names = NULL
)
}
\arguments{
\item{df_adjacency}{A square matrix or data frame or a list of these. The data
should reprensed the strength of the relation between what will be the nodes in
the network. Rownames and column names are required.}
}
\value{
The section on the returned values
}
\description{
This function creates a visualization for the input network
information. The present implementation creates a network in which nodes are
arranged in a circle. To highlight relevant interactions, there are different
options available that determine the scaling of the weights in the network.
}
\section{Additional criteria for the use of this function}{

\itemize{
\item The Cytoscape software needs to be running.
\item ...
}
}

\examples{
sum(1:10)

}
